Puppeteer tests: Convert fetchPage to page export (#2363).
Continuation of #2440.
