+++ src/e2e/puppeteer/helpers/asBrowserEnvironment.ts
@@ -1,10 +1,10 @@
 import { BrowserEnvironment } from '../../browserEnvironment/types'
-import { fetchPage } from './setup'
+import { page } from '../setup'
 
 /** Converts a Page to a BrowserEnvironment. */
 const asBrowserEnvironment = (): BrowserEnvironment => ({
   // assert UnwrapPromiseLike<R> to Promise<R>
-  execute: <R>(f: () => R) => fetchPage().evaluate(f) as Promise<R>,
+  execute: <R>(f: () => R) => page.evaluate(f) as Promise<R>,
 })
 
 export default asBrowserEnvironment
