+++ src/e2e/puppeteer/helpers/down.ts
@@ -1,9 +1,9 @@
 import { KeyInput, Keyboard } from 'puppeteer'
-import { fetchPage } from './setup'
+import { page } from '../setup'
 
 type Options = Parameters<Keyboard['down']>[1]
 
 /** Holds down a key on the keyboad. */
-const down = (key: KeyInput, options?: Options) => fetchPage().keyboard.down(key, options)
+const down = (key: KeyInput, options?: Options) => page.keyboard.down(key, options)
 
 export default down
