+++ src/e2e/puppeteer/helpers/emulate.ts
@@ -1,7 +1,7 @@
 import { Device } from 'puppeteer'
-import { fetchPage } from './setup'
+import { page } from '../setup'
 
 /** Holds down a key on the keyboad. */
-const emulate = async (device: Device) => await fetchPage().emulate(device)
+const emulate = async (device: Device) => await page.emulate(device)
 
 export default emulate
