+++ src/e2e/puppeteer/helpers/getComputedColor.ts
@@ -1,11 +1,11 @@
 import { ElementHandle } from 'puppeteer'
-import { fetchPage } from './setup'
+import { page } from '../setup'
 
 /**
  * Get computed color.
  */
 const getComputedColor = async (element: ElementHandle) => {
-  const styleHandle = await fetchPage().evaluateHandle(elementHandle => {
+  const styleHandle = await page.evaluateHandle(elementHandle => {
     const cssDeclarationObject = window.getComputedStyle(elementHandle)
     return cssDeclarationObject.color
   }, element)
