+++ src/e2e/puppeteer/helpers/getEditable.ts
@@ -1,10 +1,10 @@
-import { fetchPage } from './setup'
+import { page } from '../setup'
 
 /**
  * Get editable node handle for the given value.
  */
 const getEditable = (value: string) =>
-  fetchPage().evaluateHandle(value => {
+  page.evaluateHandle(value => {
     const xpath = `//div[@data-editable and contains(text(), "${value}")]`
     return document.evaluate(xpath, document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null)
       .singleNodeValue as HTMLElement
