+++ src/e2e/puppeteer/helpers/newThought.ts
@@ -1,9 +1,8 @@
-import { fetchPage } from './setup'
+import { page } from '../setup'
 import waitForEditable from './waitForEditable'
 
 /** Creates a new thought by hitting Enter and typing text. Waits for renders between each step. */
 const newThought = async (value?: string) => {
-  const page = fetchPage()
   await page.keyboard.press('Enter')
   await waitForEditable('')
   if (value) {
