+++ src/e2e/puppeteer/helpers/paste.ts
@@ -1,6 +1,6 @@
 import { HOME_TOKEN } from '../../../constants'
 import { WindowEm } from '../../../initialize'
-import { fetchPage } from './setup'
+import { page } from '../setup'
 
 const em = window.em as WindowEm
 
@@ -9,7 +9,6 @@ async function paste(pathUnranked: string[], text: string): Promise<void>
 
 /** Import text on given unranked path using exposed testHelpers. */
 async function paste(pathUnranked: string | string[], text?: string): Promise<void> {
-  const page = fetchPage()
   const _pathUnranked = typeof pathUnranked === 'string' ? [HOME_TOKEN] : (pathUnranked as string[])
   const _text = typeof pathUnranked === 'string' ? pathUnranked : text!
   // await new Promise(resolve => setTimeout(resolve, 200))
