+++ src/e2e/puppeteer/helpers/press.ts
@@ -1,9 +1,9 @@
 import { KeyInput, Keyboard } from 'puppeteer'
-import { fetchPage } from './setup'
+import { page } from '../setup'
 
 type Options = Parameters<Keyboard['press']>[1]
 
 /** Presses a key on the keyboad. */
-const press = (key: KeyInput, options?: Options) => fetchPage().keyboard.press(key, options)
+const press = (key: KeyInput, options?: Options) => page.keyboard.press(key, options)
 
 export default press
