+++ src/e2e/puppeteer/helpers/refresh.ts
@@ -1,6 +1,6 @@
-import { fetchPage } from './setup'
+import { page } from '../setup'
 
 /** Refreshes the page. */
-const refresh = () => fetchPage().evaluate(() => window.location.reload())
+const refresh = () => page.evaluate(() => window.location.reload())
 
 export default refresh
