+++ src/e2e/puppeteer/helpers/remove.ts
@@ -1,8 +1,8 @@
-import { fetchPage } from './setup'
+import { page } from '../setup'
 
 /** Removes the first Node that matches the selector from the DOM. NOOP if the selector is empty. */
 const remove = async (selector: string) => {
-  return fetchPage().evaluate((selector: string) => document.querySelector(selector)?.remove(), selector)
+  return page.evaluate((selector: string) => document.querySelector(selector)?.remove(), selector)
 }
 
 export default remove
