+++ src/e2e/puppeteer/helpers/scroll.ts
@@ -1,8 +1,8 @@
-import { fetchPage } from './setup'
+import { page } from '../setup'
 
 /** Scroll to the top of the page. */
 const scroll = async (x: number, y: number) => {
-  await fetchPage().evaluate(
+  await page.evaluate(
     (x: number, y: number) => {
       window.scroll(x, y)
     },
