+++ src/e2e/puppeteer/helpers/waitForHiddenEditable.ts
@@ -1,11 +1,11 @@
-import { fetchPage } from './setup'
+import { page } from '../setup'
 import waitForEditable from './waitForEditable'
 
 /** Wait for an editable to become hidden by checking the color alpha. */
 // TODO: Broken after virtualizing thoughts
 async function waitForHiddenEditable(value: string) {
   const editableElement = await waitForEditable(value)
-  await fetchPage().waitForFunction(
+  await page.waitForFunction(
     (element: Element) => {
       return window.getComputedStyle(element, null).color === 'rgba(255, 255, 255, 0)'
     },
