+++ src/e2e/puppeteer/helpers/waitForState.ts
@@ -1,5 +1,5 @@
 import { WindowEm } from '../../../initialize'
-import { fetchPage } from './setup'
+import { page } from '../setup'
 
 const em = window.em as WindowEm
 
@@ -7,7 +7,7 @@ const em = window.em as WindowEm
  * Wait until value of the state for the given property path equals the given value.
  */
 const waitForState = async (path: string, value: any) => {
-  await fetchPage().evaluate(
+  await page.evaluate(
     async (path, value) => {
       await new Promise(resolve => {
         const { getState, _ } = em.testHelpers
