+++ src/e2e/puppeteer/helpers/waitUntil.ts
@@ -1,6 +1,6 @@
-import { fetchPage } from './setup'
+import { page } from '../setup'
 
 /** Executes a function in the page context at an interval until it returns truthy. */
-const waitUntil = <R>(f: () => R) => fetchPage().waitForFunction(f)
+const waitUntil = <R>(f: () => R) => page.waitForFunction(f)
 
 export default waitUntil
